requirements.txt
selenium
webdriver-manager
python-dotenv  # Only if you use .env
